+++ src/TwigContextInterface.php
@@ -1,4 +1,5 @@
-<?php declare(strict_types=1);
+<?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
@@ -6,14 +7,10 @@
 
 namespace OxidEsales\Twig;
 
-/**
- * Interface TwigContextInterface
- *
- * @package OxidEsales\Twig
- */
 interface TwigContextInterface
 {
     /**
+     * @deprecated will be removed in v2.0.
      * @return array
      */
     public function getTemplateDirectories(): array;
@@ -24,6 +21,8 @@ interface TwigContextInterface
     public function getIsDebug(): bool;
 
     /**
+     * @deprecated will be removed in v2.0.
+     * Use \OxidEsales\EshopCommunity\Internal\Transition\Utility\BasicContextInterface::getTemplateCacheDirectory()
      * @return string
      */
     public function getCacheDir(): string;