+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - Unreleased
+
+### Deprecated
+- Method:
+    - `TwigContextInterface::`
+      - `getCacheDir()`
+      - `getTemplateDirectories()`
+
 ## [1.1.0] - 2022-09-08
 
 ### Deprecated